✨ feat: 修改更新用户方式

+++ static/js/common/userSetting.js
@@ -1,22 +1,38 @@
 layui.use(["form", "miniTab"], function () {
 	var form = layui.form,
 		layer = layui.layer,
-		miniTab = layui.miniTab;
+		miniTab = layui.miniTab,
+		$ = layui.$;
 
 	//监听提交
 	form.on("submit(saveBtn)", function (data) {
-		var index = layer.alert(
-			JSON.stringify(data.field),
-			{
-				title: "最终的提交信息",
-			},
-			function () {
-				layer.close(index);
-				miniTab.deleteCurrentByIframe();
-			}
-		);
-		return false;
+		layer.confirm('确认修改?', function(){
+			console.log(data.field);
+			$.ajax({
+				type : "post",
+				url : "user_setting.json?action=changeSetting",
+				data : data.field,
+				success : function(res){
+					if (res.code == 0) {
+						layer.msg('修改信息成功');
+						miniTab.deleteCurrentByIframe();
+					} else {
+						layer.msg(res.msg);
+					}
+				},
+			});
+		});
 	});
 
 	form.val("baseinfo", userInfo);
+
+	form.verify({
+		verifyEmail: function (value) {
+			var ePattern =
+				/^([A-Za-z0-9_\-\.])+\@([A-Za-z0-9_\-\.])+\.([A-Za-z]{2,4})$/;
+			if (value != "" && !ePattern.test(value)) {
+				return "邮箱格式不正确";
+			}
+		},
+	});
 });
